cleanup (NewsCarousel): move to `~/components/Base`
