+++ components/Base/NewsCarousel/NewsCarousel.vue
