+++ components/Base/NewsCarousel/index.js
